[MH3-17812] - used proper pointer cast

+++ apps/tcpip_tcp_client_server/firmware/src/config/sam_a5d2_xult/library/tcpip/src/tcpip_announce.c
@@ -353,7 +353,7 @@ bool TCPIP_ANNOUNCE_Initialize(const TCPIP_STACK_MODULE_CTRL* const stackCtrl, c
         TCPIP_UDP_OptionsGet(s, UDP_OPTION_TX_BUFF, (void*)&currBuffSize);
         if(currBuffSize < announceDcpt.txBuffSize)
         {
-            if(!TCPIP_UDP_OptionsSet(s, UDP_OPTION_TX_BUFF, (void*)announceDcpt.txBuffSize))
+            if(!TCPIP_UDP_OptionsSet(s, UDP_OPTION_TX_BUFF, (void*)(uintptr_t)announceDcpt.txBuffSize))
             {
                 break;
             }
